Added account settings, infinite scrolling and collections page

+++ server/controllers/auth/user.controller.js
@@ -1,8 +1,9 @@
 import Joi from "joi";
-import { User } from "../../models/index.js";
+import { Collection, Post, SavedPost, User } from "../../models/index.js";
 import CustomErrorHandler from "../../services/CustomErrorHandler.js";
 import bcrypt from "bcrypt";
 import fs from "fs";
+import cloudinary from "../../config/cloudinary.js";
 
 const userController = {
   async me(req, res, next) {
@@ -27,7 +28,7 @@ const userController = {
       let user = await User.findById(userId);
 
       // remove current avatar from folder
-      if (user.avatar)
+      if (user.avatar && user.avatar.startsWith("uploads"))
         fs.unlinkSync(`${appRoot}/${user.avatar}`, (err) => {
           return next(err);
         });
@@ -48,9 +49,12 @@ const userController = {
       const { _id: userId } = req.user;
       let user = await User.findById(userId);
 
-      fs.unlinkSync(`${appRoot}/${user.avatar}`, (err) => {
-        return next(err);
-      });
+      if (user.avatar && user.avatar.startsWith("uploads")) {
+        fs.unlinkSync(`${appRoot}/${user.avatar}`, (err) => {
+          return next(err);
+        });
+      }
+
       user = await User.findByIdAndUpdate(
         userId,
         { avatar: null },
@@ -98,6 +102,51 @@ const userController = {
         message: "Password changed successfully. Please login again..!",
       });
     } catch (err) {}
+  },
+
+  async deactivateUser(req, res, next) {
+    try {
+      await User.findByIdAndUpdate(req.user?._id, { deactivated: true });
+      return res
+        .status(200)
+        .json({ success: true, message: "Account deactivated successfully" });
+    } catch (err) {
+      return next(err);
+    }
+  },
+
+  async deleteUser(req, res, next) {
+    try {
+      const { _id: userId } = req?.user;
+
+      // deleting userPost from cloudinary
+      const delPosts = await Post.find({ user: userId });
+      delPosts.forEach(async (node) => {
+        await cloudinary.uploader.destroy(node.image.id);
+      });
+
+      // deleting userPosts from database
+      await Post.deleteMany({ user: userId });
+      await Collection.deleteMany({ collectionId: userId });
+
+      // deletings userSavedPosts from cloudinary
+      const delSavedPosts = await SavedPost.find({ user: userId });
+      delSavedPosts.forEach(async (node) => {
+        await cloudinary.uploader.destroy(node.image.id);
+      });
+
+      // deletings userSavedPosts from databse
+      await SavedPost.deleteMany({ user: userId });
+
+      // Finally deleting user
+      await User.findByIdAndDelete(userId);
+
+      return res
+        .status(200)
+        .json({ success: true, message: "Account deleted successfully" });
+    } catch (err) {
+      return next(err);
+    }
   },
 };
